test(static): harden pr-bot degraded local-fallback against Static flake
Under the exact Static partition, execute_pr_bot_degraded_local_fallback_records_rationale_and_reaches_merge
failed once with local-fallback launch rc=1 while focused and pr-bot suite stress stayed green.

- Install native-review-bypass.sh in the degraded gate fixture so CSA_WORKFLOW_DIR helpers are complete.
- Split degraded fixtures under the monolith token budget.
- Serialize the flake test via nextest threads-required (retries remain 0) (#2847).
